% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommaValues.R
\name{commas}
\alias{commas}
\title{Format large values to include comma separators and rounding}
\usage{
commas(x, scinot = FALSE)
}
\arguments{
\item{x}{numerical value for which to add comma formatting to}

\item{scinot}{logical value to permit scientific notation in output value. Default is _FALSE_}
}
\value{
A numerical value rounded to the nearest 2 decimal places
}
\description{
Format large values to include comma separators and rounding
}
\examples{
commas(12345610)
commas(12345610, TRUE)
commas(0.123456)
}
