% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFolders.R
\name{create_proj_dir}
\alias{create_proj_dir}
\title{Create Project Directory}
\usage{
create_proj_dir(extra = NULL)
}
\arguments{
\item{extra}{Insert extra folders as required as a character vector}
}
\value{
A project folder structure
}
\description{
Create Project Directory
}
\examples{
create_proj_dir()
create_proj_dir(c("Stuff", "More Stuff", "info", "info/dat"))

}
