% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cardinal direction.R
\name{insert_cardinal_dir}
\alias{insert_cardinal_dir}
\title{Insert the cardinal compass direction for given decmimal degress}
\usage{
insert_cardinal_dir(data, dir_col)
}
\arguments{
\item{data}{a dataframe containing the data}

\item{!!dir_col}{direction column to convert to cardinal factor.}
}
\value{
the original dataframe with the cardinal varaible included
}
\description{
This function inserts a new variable "cardinal.dir" into the dataframe passed to it..
}
\examples{

insert_cardinal_dir(waves, dir_deg)

#use with dply
library(dplyr)
waves \%>\% insert_cardinal_dir(waves, dir_deg)
}
