% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Points Possible Check.R
\name{points_possible_check}
\alias{points_possible_check}
\title{Checking if total points row is present}
\usage{
points_possible_check(gradebook)
}
\arguments{
\item{gradebook}{A gradebook file from Canvas}
}
\value{
Outputs True/False value
}
\description{
Internal use function to test whether total points possible row is present in a given gradebook input
}

