% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmootheCDF.R
\name{smooth_ecd}
\alias{smooth_ecd}
\title{Create a smoothed (theoretical) Cumulative Distribution Function plot}
\usage{
smooth_ecd(
  data,
  y,
  adj = 1,
  color_var = NULL,
  title = NULL,
  subtitle = NULL,
  compare = FALSE,
  ylab = NULL,
  coord_flip = FALSE,
  plot_only = TRUE,
  show_percent = FALSE
)
}
\arguments{
\item{data}{a dataframe containing the data.}

\item{y}{column reference to determine exceedance entered as an object (without quotes).}

\item{adj}{adjustment value for smoothing. Select values between 0 and 1.}

\item{color_var}{\emph{optional} a grouping variable entered as an object reference.}

\item{title}{\emph{optional} title for resulting plot.}

\item{subtitle}{\emph{optional} a subtitle for the resulting plot if required.}

\item{compare}{\emph{optional} logical value to output the true Empirical Cumulative Distribution Function with the smoothed function, providing a visual comparison.}

\item{coord_flip}{\emph{optional} to flip the X and Y axis for visual preference.}

\item{plot_only}{\emph{optional} when \code{TRUE} a ggtable object is returned without appending the data.}

\item{show_percent}{\emph{optional} when \code{TRUE} percentage labels are shown on the y axis instead of probability.}
}
\value{
Depending on the argument passed to \code{plot_only} either a list object is returned when \code{plot_only == FALSE} containing a grob object \code{[["plot"]]} for plotting and a dataframe \code{[["data"]]} for further analysis
. A single plot is returned when \code{plot_only == TRUE}.
}
\description{
This function produces a smoothed (theoretical) eCDF plot.
}
\examples{
data(waves)

 plt_a <- smooth_ecd(data = waves,
 y = height_meters)

 plot(plt_a)

 plt_a <- smooth_ecd(data = waves,
 y = height_meters,
 plot_only = FALSE)

 plot(plt_a[["plot"]])

 # Add a grouping variable
 library(dplyr)

 plt_b <- smooth_ecd(data = waves \%>\%
 mutate(hour = format(date_time, "\%H")),
 y = height_meters,
 color_var = hour,
 adj = 1,
 title = "Wave Height Cumulative Distribution",
 subtitle = "By Hour",
 plot_only = FALSE)

 # Access the list objects
 plot(plt_b[["plot"]])
 
 dat <- plt_b[["data"]]

}
