% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical ranges.R
\name{cat_num}
\alias{cat_num}
\title{Insert the categorical label for a vector of numerical values.}
\usage{
cat_num(data, num_col, bins = 4)
}
\arguments{
\item{data}{a dataframe containing the data}

\item{num_col}{is the column name containing the numerical values to be binned and labeled.}

\item{bins}{is the number of bins to assign the categorical labels}
}
\value{
the categorical labels as a factor
}
\description{
Numerical data is passed along with the number of bins required. A label of binned factor ranges is returned
}
\examples{

# Create some data
samples <- data.frame(samp = sample(1:1000, 500, replace = T))

cat_num(samples, samp)

# Use with dplyr

library(dplyr)
samples \%>\% cat_num(samp)
}
