% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmootheCDF.R
\name{smooth_ecd}
\alias{smooth_ecd}
\title{Create a smoothed Cumulative Distribution Function plot}
\usage{
smooth_ecd(
  data,
  y,
  adj = 1,
  color_var = NULL,
  title = NULL,
  subtitle = NULL,
  compare = FALSE
)
}
\arguments{
\item{data}{a dataframe containing the data}

\item{y}{column reference to determine exceedance entered as an object (without quotes).}

\item{adj}{adjustment value for smoothing. Select values between 0 and 1.}

\item{color_var}{\emph{optional} a grouping variable entered as an object reference.}

\item{title}{\emph{optional} title for resulting plot.}

\item{subtitle}{\emph{optional} a subtitle for the resulting plot if required.}

\item{compare}{\emph{optional} logical value to show ggplot's stat_ecdf geom to provide a visual cue and confidence.}
}
\value{
A list object is returned containing a grob object [["plot"]] for plotting a dataframe [["data"]] for further analysis
}
\description{
This function produces a smoothed eCDF plot.
}
\examples{
data(waves)

plt_a <- smooth_ecd(data = waves,
 compare = TRUE,
 y = height_meters,
 adj = 1)

 plot(plt_a[["plot"]])

 ## Add a grouping variable
 library(dplyr)

 plt_b <- smooth_ecd(data = waves \%>\%
 mutate(hour = format(date_time, "\%H")),
 y = height_meters,
 color_var = hour,
 adj = 1,
 title = "Wave Height Exceedance",
 subtitle = "By Hour")

 # Access the objects
 plot(plt_b[["plot"]])
 dat <- plt_b[["data"]]

}
